TAR-26 Changed throttling configuration value from float (0.00 to 1) to int (0 to 100)

+++ app/model/AuditContext.scala
@@ -123,4 +123,4 @@ trait TAuditContext {
 
 case class AuditContext() extends TAuditContext
 
-case class ThrottlingAuditContext(throttlingPercentage: Option[Float], throttled: Boolean, initialDestination: LocationType, throttlingEnabled: Boolean)
+case class ThrottlingAuditContext(throttlingPercentage: Option[Int], throttled: Boolean, initialDestination: LocationType, throttlingEnabled: Boolean)
